Extract setup provisioning service
