+++ mhm/src-tauri/src/services/setup/mod.rs
@@ -1,5 +1,7 @@
-pub mod status;
+mod provisioning;
+mod status;
 
+pub use provisioning::complete_setup;
 pub use status::read_bootstrap_status;
 
 #[cfg(test)]
